refactor: replace Navbar component with LandingPage and create Layout component
