+++ imagefic/src/Components/Layout.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Layout = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Layout
